requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
chromadb==0.5.0
python-multipart==0.0.20
requests==2.32.3
jinja2==3.1.4
python-dotenv==1.0.1
